docs: split Events into flat per-listener pages, smooth sidebar collapse, full-bleed sidebar
Events
- One page per listener under /docs/events/{purchase-updated-listener,
  purchase-error-listener, subscription-billing-issue-listener,
  ios/promoted-product-listener-ios,
  android/user-choice-billing-listener-android,
  android/developer-provided-billing-listener-android}, mirroring the
  flat APIs/Types layout
- /docs/events becomes the index/hub with Event System Overview,
  three listener tables, and Event Listener Management
- Events sidebar entry is now a MenuDropdown with iOS Specific /
  Android Specific subgroups
- Every cross-doc link now routes straight to its listener page —
  home.tsx chips, introduction.tsx, fetch-products / request-purchase
  Important callouts, lifecycle/subscription, etc.

Sidebar
- Collapse animation reworked as CSS-only grid-template-rows: 0fr → 1fr,
  no more JS height measurement (which clipped iOS/Android Specific
  contents to a single row)
- New --sidebar-edge-fill variable + negative left margin extends the
  sidebar background and active-row highlight to the viewport edge
- Nested group headers (iOS Specific / Android Specific) indent one
  level deeper than the listener rows above them so the hierarchy
  reads correctly
- Drop the "Documentation" h3 (the Docs route already labels the section)
  and trim the top padding above Getting Started
