+++ packages/docs/src/pages/docs/apis/fetch-products.tsx
@@ -33,11 +33,11 @@ function FetchProducts() {
             not rely on their return values for actual purchase results
           </strong>
           . Instead, listen for events through{' '}
-          <Link to="/docs/events#purchase-updated-listener">
+          <Link to="/docs/events/purchase-updated-listener">
             <code>purchaseUpdatedListener</code>
           </Link>{' '}
           or{' '}
-          <Link to="/docs/events#purchase-error-listener">
+          <Link to="/docs/events/purchase-error-listener">
             <code>purchaseErrorListener</code>
           </Link>
           .
